Ignore mime-type if Active Agency User Uploads File
Due to inconsistencies in the way agencies generate some files,
if an active agency user uploads a file, then the file is automatically,
considered to be valid by the system.

Files will still be scanned by the virus scanner, if enabled, to
prevent viruses from unintentionally being uploaded to the system.

This only ignores the validity of the mime-type generated by the magic
file and the operating system.