Only create a new user if not an agency user.
When a public user (anonymous or NYC.ID) logs in with a new account
(defined as a new GUID + Auth Type) a new account will be created,
regardless of the email address on the account.
Agency users will be modified if their user account is using the custom
LDAP user type.