Remove response_token Expiration Date.
Removed the response_token.expiration_date from the ResponseTokens
class. Tokens no longer expire. If a file is released and public,
then the token will be checked before the release date of the file,
otherwise it is ignored. If a file is released and private, then a
token will always be checked.