Feature Flag for Monitoring of Sub Agencies
This change adds a new field to the Agencies table to determine if an agency should monitor it's sub agencies requests.

A sub agency is an agency that shares it's EIN with another agency. An example of this is the Mayor's Office (Parent Agency) who shares its EIN with the Mayor's Office of Criminal Justice. In this case, the agency and the sub agency will have different ein and parent_eins. Internally, requests are routed to the right agency, but externally, the Parent EIN is used to generate the request ID.

For agencies with sub agencies, the current behavior is to always add the administrators from the parent agency to a new request. As per a request from the Mayor's Office, we have moved this to a feature flag that can be turned on if the agency so chooses. It will default to not providing this behavior, as it is rarely used.

A migration has been added to include this change. The comments on the Agencies model have also been updated to reflect all of the fields that are currently in place.

Signed-off-by: Joel Castillo <jcastillo@records.nyc.gov>