Add JSON Field for Agency Features Class
This commit is the initial version to add a JSON field to the agencies table to
handle the optional features in OpenRecords. This initial version contains the
feature flag for the agencies that want to follow the requests of their sub
agencies.

This is not complete. Unit tests still need to be added as well as more
complete manual testing