Add Character Limit for File Titles
This modification adds a limit of 140 characters to the title of any file that is uploaded via the 'Add Files' upload dialog.

There is front-end validation that prevents the user from moving forward in the submission form.

This commit includes a database migration to change the database character limit to 140 characters.

Signed-off-by: Joel Castillo <jcastillo@records.nyc.gov>