Add Style to Character Count
Added style to character counter to identify when the character count of the specified field is below the minimum length.

The function now has a default parameter (with value of 0) for the minimum length. If the length is below or above the character count limits, the character counter will be red. Otherwise, it is black.

Signed-off-by: Joel Castillo <jcastillo@records.nyc.gov>