bugfix/OP-1303: Add Link to Files Submitted with Request
Add a file link to the email confirmation / receipt sent out to the agency and requester when a request is submitted on OpenRecords. This only applies when a file has been added to the request. The link contains a token for the release and private file that the requester has provided as part of their request.

The link is the same for agencies that are live on the portal and agencies that have not begun using the portal to respond to FOIL requests.

Signed-off-by: Joel Castillo <jocastillo@records.nyc.gov>