Add Error Handling for Anonymous Users and KeyErrors
Added a check for a KeyError when trying to delete a session. If the
session does not exist, Redis will raise a KeyError. This means we don't
need to destroy the session and can just move on.

Additionally, if the current_user is anonymous, we don't need to do any
database modifications, because the user does not exist in the Users
table. THerefore we can just redirect to the home page when they click
logout.

Signed-off-by: Joel Castillo <jocastillo@records.nyc.gov>