Add Duplicate File Check in Files class
Added a new exception `DuplicateFileException` which is raised when a
file is uploaded to a request a second time. This is used in a new check
in the constructor for the Files class in models.py to ensure that a
file is not uploaded twice. This is done to prevent a second email from
being sent out to the requester when a file is uploaded twice due to a
race condition with the Redis keystore.

If a file is uploaded twice, the upload will go through with all valid
files, but any file that was not successfully uploaded will display an
alert at the top of the page after the form has been submitted.

Finally, the old models from legacy v2.0 has been removed from the app
folder, and the new models with the appropriate migrations have been
put in place.

Signed-off-by: Joel Castillo <jocastillo@records.nyc.gov>