Fix Multiple Emails for Multi-File Uploads
Due to a misplaced function call inside of a for loop, when a user uploads multiple files, one email per file was being sent out in the format:
Email 1: File A
Email 2: File A, File B
Email 3: File A, File B, File C
etc...

To resolve this issue, the following changes were made to the `send_file_email` function, which generates the emails for the file upload utility.
Email Generation Conditions
 - "Release and Public" links provided
 - "Release and Private" links provided
 - "Private" links provided
 - "Release and Public" links and "Release and Private" links provided
 - "Release and Public" links and "Private" links provided
 - "Release and Private" links and "Private" links provided
 - "Release and Public" links and "Release and Private" links and "Private" links provided.

Previously even if no links were provided, an email without links could be sent due to an else condition. This condition has been changed to an elif.

Signed-off-by: Joel Castillo <jocastillo@records.nyc.gov>