Implement Maintenance Pages for OpenRecords
This commit adds maintenance pages to OpenRecords using two methods.

1) Using Nginx:
A variation of this article (No nginx basic auth with either network or cookie set)[https://www.liip.ch/en/blog/no-nginx-basic-auth-with-either-network-or-cookie-set] allows us to check to see if the user is accessing OpenRecords from either an internal (approved) network or from a computer with the appropriate cookie set. If so, the user can go directly to OpenRecords. If not, the user is sent a 503 message and told that the site is down for maintenance.

2) Using Flask
Using the app.instance_path variable, we check before each request for the existance of the "maintenance.json" file in the instance folder. If it exists and the authorized_maintainer cookie does not exist, the user is sent to a 503 page (using the abort method). The page pulls data from the maintenance.json file (description and outage_time) to display a detailed message to the user during maintenance.
Users who have the appropriate cookie set can continue to use the application.

To set cookies, use the following code snippet:
`document.cookie = 'authorized_maintainer=true;expires=<TIME IN MILLISECONDS>;path=/'`

You can also use an extension to add the cooke (EditThisCookie for Chrome works well).

Signed-off-by: Joel Castillo <jocastillo@records.nyc.gov>