Add CLI flag for updating Events table
Adds a CLI flag for updating the events table to include
request_status_changed events for requests closed between the current
date and the release of v2.3.

Additionally update the Events table to resolve an incorrect key for
request status changes (jobs.py) and use the REQ_DENIED event
instead of REQ_CLOSED for denials

Finally, adds a request_status_changed event for acknowledging a
request.