Resolve error with state and zip code.
Resolves the issue where the state was being set to None if the modal
dialog was closed before hitting submit. This was due to the way select
fields are handled. In order to get the value of a select field you need
to select the option value within the select field and get its value.
Without this, an empty string was being returned.
To resolve this, the state value, if present, is stored in JS on page
load and then used on the reset modal callback to restore the state.

To resolve the issue with the missing state and zipcode on the backend,
the UserAddressDict was modified to use the right keys when storing
the values of zip and state (in the case of state, it just wasn't
present in the initialization function.)