Check for the presence of SSN in Form Submission
Adds a function in main.js (checkSSN) to look through the provided
string and match the regex for a possible social security number.

The validation on form submission will check the title, description,
and / or custom request forms content for the presence of a matching
string.

If a string match is found, a modal dialog confirmation will appear
giving the user the option to either submit the form or review their
submission to edit any possible PII.