Update show_title to use status instead of last_date_closed
Issue with the last_date_closed is due to an agency go-live after we added the date_closed column to the requests table.

Additionally resolves and issue with the Holiday bank to generate holidays back to the beginning of the portal. We may need to optimize this later on as the portal ages.